backend/pyproject.toml
[project]
name = "interactive-story-backend"
version = "1.0.0"
description = "Interactive Story Agent — Google ADK + A2A + A2UI"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "a2a-sdk>=0.2.0",
    "uvicorn[standard]>=0.30.0",
    "starlette>=0.37.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
story-agent = "story_agent.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["story_agent"]
